roll worker file into main indexdocs
